incorporate Component.js and update sign in page
