--- frontend/pages/SignIn.js
@@ -1,39 +1,0 @@
-class SignIn extends HTMLElement {
-    constructor() {
-      super();
-      this.attachShadow({ mode: 'open' });
-    }
-  
-    connectedCallback() {
-      this.render();
-    }
-  
-    render() {
-      if (this.isUserAuthenticated()) {
-        this.redirectToHome();
-        return;
-      }
-  
-      this.shadowRoot.innerHTML = `
-        <style>
-          /* Add styles here */
-        </style>
-        <navbar-component></navbar-component>
-        <sign-in-content></sign-in-content>
-      `;
-    }
-  
-    isUserAuthenticated() {
-      // Placeholder for actual authentication logic
-      return window.userManagementClient && window.userManagementClient.isAuth();
-    }
-  
-    redirectToHome() {
-      // Placeholder for redirect logic
-      window.getRouter().redirect('/');
-    }
-  }
-  
-  // Define the custom element
-  customElements.define('sign-in', SignIn);
-
