+++ frontend/pages/SignInContent.js
@@ -1,53 +1,64 @@
-class SignInContent extends HTMLElement {
-    constructor() {
-      super();
-      this.attachShadow({ mode: 'open' });
-      this.isValidEmailInput = false;
-      this.isValidPasswordInput = false;
-    }
-  
-    connectedCallback() {
-      this.render();
-    }
-  
-    render() {
-      this.shadowRoot.innerHTML = `
-        <style>
-          /* Add styles here */
-        </style>
-        <div>
-          <h2>Sign In</h2>
-          <form id="signin-form">
-            <label for="email">Email:</label>
-            <input type="email" id="email" name="email" required>
-            <label for="password">Password:</label>
-            <input type="password" id="password" name="password" required>
-            <button type="submit">Sign In</button>
-          </form>
-          <div id="error-message"></div>
-        </div>
-      `;
-  
-      this.shadowRoot.querySelector('#signin-form').addEventListener('submit', (event) => {
-        event.preventDefault();
-        this.handleSignIn();
-      });
-    }
-  
-    handleSignIn() {
-      const email = this.shadowRoot.querySelector('#email').value;
-      const password = this.shadowRoot.querySelector('#password').value;
-  
-      // Perform input validation and API call here
-      if (email && password) {
-        // Call to user management client or API
-        console.log('Sign-in request:', { email, password });
-      } else {
-        this.shadowRoot.querySelector('#error-message').textContent = 'Please fill out all fields.';
-      }
-    }
-  }
-  
-  // Define the custom element
-  customElements.define('sign-in-content', SignInContent);
-  
+import { login } from "../js/clients/token-client.js";
+
+export class SignInContent extends Component {
+	constructor() {
+		super();
+		this.isValidEmailInput = false;
+		this.isValidPasswordInput = false;
+		this.passwordHiden = true;
+
+		this.error = false;
+		this.errorMessage = "";
+	}
+
+	render() {
+		// const { render } = this.#OAuthReturn();
+		// if (!render) {
+		// 	return false;
+		// }
+		this.innerHTML = `
+
+    `;
+	}
+
+
+
+	// async #loadAndCache(refreshToken) {
+	// 	this.innerHTML = this.#renderLoader();
+	// 	userManagementClient.refreshToken = refreshToken;
+	// 	if (!(await userManagementClient.restoreCache())) {
+	// 		userManagementClient.logout();
+	// 		this.error = true;
+	// 		this.errorMessage = "Error, failed to store cache";
+	// 		this.reRender();
+	// 	} else {
+	// 		getRouter().navigate("/");
+	// 	}
+	// }
+
+	#togglePasswordVisibility() {
+		if (this.passwordHiden) {
+			this.password.setAttribute("type", "text");
+		} else {
+			this.password.setAttribute("type", "password");
+		}
+		this.passwordEyeIcon.children[0].classList.toggle("bi-eye-fill");
+		this.passwordEyeIcon.children[0].classList.toggle("bi-eye-slash-fill");
+		this.passwordHiden = !this.passwordHiden;
+	}
+
+	#startLoadButton() {
+		this.signinBtn.innerHTML = `
+      <span class="spinner-border spinner-border-sm" role="status" aria-hidden="true"></span>
+      <span class="sr-only">Loading...</span>
+    `;
+		this.signinBtn.disabled = true;
+	}
+
+	#resetLoadButton() {
+		this.signinBtn.innerHTML = "Sign in";
+		this.signinBtn.disabled = false;
+	}
+}
+
+customElements.define("signin-content", SignInContent);
